// _citybox_src/services/payment-api/src/generated/prisma/enums.ts
export const ProviderType = {
  ASAAS: 'ASAAS',
  PAGBANK: 'PAGBANK',
  INFINITE_PAY: 'INFINITE_PAY',
  STONE: 'STONE',
  STUB: 'STUB'
} as const

export type ProviderType = (typeof ProviderType)[keyof typeof ProviderType]